refactor rnafusion models
