+++ cg/meta/workflow/nf_analysis.py
@@ -12,7 +12,7 @@ from cg.io.controller import ReadFile, WriteFile
 from cg.meta.workflow.analysis import AnalysisAPI
 from cg.meta.workflow.nf_handlers import NextflowHandler, NfBaseHandler, NfTowerHandler
 from cg.models.cg_config import CGConfig
-from cg.models.rnafusion.arguments import CommandArgs
+from cg.models.workflow.rnafusion import CommandArgs
 from cg.utils import Process
 
 LOG = logging.getLogger(__name__)
